Wholesale changes to install - includes building Test Suite. ->
Without going into details - changed package.json to run a
postinstall script - ./test/install.js. That script installs
the Test Suite as well as changes the node .html files to
display the current version of the alasql nodes within the
node-red editor (node version set from package.json - 'version').

Documentation changed to reflect the improved user experience.
