+++ node-red-contrib-alasql.html
@@ -16,7 +16,8 @@
 </script>
 
 <script type="text/x-red" data-help-name="alasql">
-    <p>A Node-RED node to use <a href="https://github.com/agershun/alasql" target="_blank">AlaSQL</a> for fast SQL based in-memory data processing for BI and ERP applications.</p>
+    <p>Version: <i>0.2.4-develop</i><br> <!-- do not change - see .test/install.js -->
+    A Node-RED node to use <a href="https://github.com/agershun/alasql" target="_blank">AlaSQL</a> for fast SQL based in-memory data processing for BI and ERP applications.</p>
     <p>Especially useful when you have (many) different sources coming with data you want to join, filter and format with good'old SQL.</p>
     <ul>
     <li>
